Move engine common config in a public package
