+++ internal/app/singularity/oci_create_linux.go
@@ -13,9 +13,9 @@ import (
 	"path/filepath"
 
 	"github.com/opencontainers/runtime-tools/generate"
-	"github.com/sylabs/singularity/internal/pkg/runtime/engine/config"
 	"github.com/sylabs/singularity/internal/pkg/runtime/engine/oci"
 	"github.com/sylabs/singularity/internal/pkg/util/starter"
+	"github.com/sylabs/singularity/pkg/runtime/engine/config"
 )
 
 // OciCreate creates a container from an OCI bundle
