+++ internal/pkg/util/starter/starter.go
@@ -14,8 +14,8 @@ import (
 	"path/filepath"
 
 	"github.com/sylabs/singularity/internal/pkg/buildcfg"
-	"github.com/sylabs/singularity/internal/pkg/runtime/engine/config"
 	"github.com/sylabs/singularity/internal/pkg/sylog"
+	"github.com/sylabs/singularity/pkg/runtime/engine/config"
 	"golang.org/x/sys/unix"
 )
